refactor modules structure.
