--- flloat/base/Interpretation.py
@@ -1,7 +1,0 @@
-"""Base class for implement semantics."""
-
-from flloat.base.hashable import Hashable
-
-
-class Interpretation(Hashable):
-    pass
