+++ flloat/base/delta.py
@@ -6,5 +6,5 @@ from flloat.semantics.pl import PLInterpretation
 class Delta(ABC):
 
     @abstractmethod
-    def delta(self, i:PLInterpretation, epsilon=False):
+    def delta(self, i: PLInterpretation, epsilon=False):
         raise NotImplementedError
