+++ flloat/base/truths.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from flloat.base.Formula import Formula, UnaryOperator, BinaryOperator
+from flloat.base.Formula import UnaryOperator, BinaryOperator
 from flloat.base.Symbols import Symbols
 
 
@@ -9,17 +9,21 @@ class Truth(ABC):
     def truth(self, *args) -> bool:
         raise NotImplementedError
 
+
 class NotTruth(UnaryOperator, Truth):
     def __init__(self, f):
         UnaryOperator.__init__(self, f)
 
     operator_symbol = Symbols.NOT.value
+
     def truth(self, *args):
         return not self.f.truth(*args)
 
 
 class AndTruth(BinaryOperator, Truth):
+
     operator_symbol = Symbols.AND.value
+
     def truth(self, *args):
         for f in self.formulas:
             if not f.truth(*args):
@@ -27,8 +31,11 @@ class AndTruth(BinaryOperator, Truth):
         return True
         # return all(f.truth(*args) for f in self.formulas)
 
+
 class OrTruth(BinaryOperator, Truth):
+
     operator_symbol = Symbols.OR.value
+
     def truth(self, *args):
         for f in self.formulas:
             if f.truth(*args):
@@ -38,8 +45,8 @@ class OrTruth(BinaryOperator, Truth):
 
 
 class ImpliesTruth(BinaryOperator, Truth):
-    operator_symbol = Symbols.IMPLIES.value
 
+    operator_symbol = Symbols.IMPLIES.value
 
     def truth(self, *args):
         # eq_formula = self._convert()
@@ -57,12 +64,13 @@ class ImpliesTruth(BinaryOperator, Truth):
 
         return truth
 
+
 class EquivalenceTruth(BinaryOperator, Truth):
+
     operator_symbol = Symbols.EQUIVALENCE.value
 
     def truth(self, *args):
         fs = self.formulas
-        N = len(fs)
         t = [f.truth(*args) for f in fs]
         # either all true or all false
         return all(t) or not any(t)
