+++ flloat/semantics/pl.py
@@ -1,13 +1,12 @@
-from typing import FrozenSet, Set, List
+from typing import FrozenSet, Set
 
-from flloat.base.Interpretation import Interpretation
-from flloat.base.Symbol import Symbol
+from flloat.base.symbols import Symbol, Interpretation
 from flloat.utils import ObjFactory, ObjConstructor
 
 
 class _PLInterpretation(Interpretation):
 
-    def __init__(self, true_propositions:FrozenSet[Symbol]):
+    def __init__(self, true_propositions: FrozenSet[Symbol]):
         super().__init__()
         self.true_propositions = frozenset(true_propositions)
 
@@ -15,7 +14,7 @@ class _PLInterpretation(Interpretation):
         # return tuple(sorted(self.true_propositions, key=lambda x: x.name))
         return self.true_propositions
 
-    def __contains__(self, item:Symbol):
+    def __contains__(self, item: Symbol):
         return item in self.true_propositions
 
     def __iter__(self):
@@ -27,6 +26,7 @@ class _PLInterpretation(Interpretation):
     def __repr__(self):
         return self.__str__()
 
+
 class PLTrueInterpretation(_PLInterpretation):
     def __init__(self):
         super().__init__(frozenset())
@@ -37,6 +37,7 @@ class PLTrueInterpretation(_PLInterpretation):
     def __contains__(self, item):
         return True
 
+
 class PLFalseInterpretation(_PLInterpretation):
     def __init__(self):
         super().__init__(frozenset())
@@ -44,14 +45,11 @@ class PLFalseInterpretation(_PLInterpretation):
     def __contains__(self, item):
         return False
 
+
 class _PLInterpretationConstructor(ObjConstructor):
-    def __call__(self, true_propositions:Set[Symbol]):
+    def __call__(self, true_propositions: Set[Symbol]):
         f_sym = frozenset(true_propositions)
         return super().__call__(f_sym)
-
-    def fromStrings(self, strings:List[str]):
-        return self(set(Symbol(s) for s in strings))
-
 
 
 plinterpretation_factory = ObjFactory(_PLInterpretation)
